Feat: Remove mock data
